src/listings-api/requirements.txt
# Location: `/src/listings-api/requirements.txt`

Flask==3.0.0
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
psycopg2-binary==2.9.9
redis==5.0.1
prometheus-client==0.19.0
gunicorn==21.2.0
python-dotenv==1.0.0
marshmallow==3.20.1
Werkzeug==3.0.1
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.7.0